requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dataclasses==0.6
distlib==0.3.3
Django==3.2.8
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
filelock==3.3.0
hiredis==2.0.0
hyperlink==21.0.0
identify==2.3.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==21.3.0
msgpack==1.0.3
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.4.0
pre-commit==2.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==21.0.0
pytz==2021.3
PyYAML==6.0
regex==2021.10.8
requests==2.27.1
semantic-version==2.8.5
service-identity==21.1.0
setuptools-rust==1.1.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.5.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0
